Refactor KoopaBlockRenderPart
